refactor(admin-bot): split hooks.ts to multiple files
