+++ apps/admin-bot/src/core/logger.ts
@@ -1,0 +1,5 @@
+import { pino } from 'pino'
+
+const root = pino()
+
+export const createLogger = (module: string) => root.child({ module })
